Migrating to LINCC Frameworks python project template.
