+++ archive/README.md
@@ -1,0 +1,10 @@
+This directory contains **archived** configuration files. 
+
+Flexcode has been migrated to use the [LINCC Frameworks Python Project Template](https://github.com/lincc-frameworks/python-project-template) which uses a pyproject.toml file for build definitions and requirements.
+
+The pyproject.toml file replaces the need for setup.py, setup.cfg, and ton.ini, 
+but the files will be retained for a period of time as a convenience. 
+
+**These files will be removed at a later date!** 
+Please do not use them. 
+Update pyproject.toml in the base directory if changes to the build are required.
