+++ docs/index.rst
@@ -1,0 +1,13 @@
+.. flexcode documentation main file.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Welcome to flexcode's documentation!
+========================================================================================
+
+.. toctree::
+   :hidden:
+
+   Home page <self>
+   API Reference <autoapi/index>
+   Notebooks <notebooks>
