+++ docs/notebooks.rst
@@ -1,0 +1,6 @@
+Notebooks
+========================================================================================
+
+.. toctree::
+
+    Introducing Jupyter Notebooks <notebooks/intro_notebook>
