+++ docs/requirements.txt
@@ -1,0 +1,9 @@
+sphinx==6.1.3
+sphinx_rtd_theme==1.2.0
+sphinx-autoapi==2.0.1
+nbsphinx
+ipython
+jupytext
+jupyter
+matplotlib
+numpy
